+++ libs/helpers/src/lib/util/rxjs/rxjs.spec.ts
@@ -60,65 +60,4 @@ describe('RXJS js operators', () => {
         complete: () => done(),
       });
   });
-
-  describe('raceEmit', () => {
-    // Тест для raceEmit: проверяем, что первый источник, который эмитирует значение в заданном временном интервале, проходит
-    it('should emit the first observable to emit within a given time frame', done => {
-      const source1 = of('first').pipe(delay(100));
-      const source2 = of('second').pipe(delay(400));
-      raceEmit(250, source1, source2).subscribe({
-        next: value => expect(value).toBe('first'),
-        complete: () => done(),
-      });
-    });
-  });
-
-  describe('moveInEventLoopIteration', () => {
-    // Тест для moveInEventLoopIteration: проверяем, что значение эмитируется без задержки, если count равен 0
-    it('should emit the value without delay if count is 0', done => {
-      const start = Date.now();
-      of('test')
-        .pipe(moveInEventLoopIteration(0))
-        .subscribe(() => {
-          const duration = Date.now() - start;
-          expect(duration).toBeLessThan(5);
-          done();
-        });
-    });
-
-    // Тест для moveInEventLoopIteration: проверяем, что значение эмитируется с задержкой, если count больше 0
-    it('should emit the value with delay if count is greater than 0', done => {
-      const start = Date.now();
-      of('test')
-        .pipe(moveInEventLoopIteration(1))
-        .subscribe(() => {
-          const duration = Date.now() - start;
-          expect(duration).toBeGreaterThanOrEqual(1);
-          done();
-        });
-    });
-
-    // Тест для moveInEventLoopIteration: проверяем, что задержка увеличивается с увеличением count
-    it('should increase delay with larger count', done => {
-      const start1 = Date.now();
-      const start2 = Date.now();
-      let duration1: number;
-      let duration2: number;
-
-      of('test')
-        .pipe(moveInEventLoopIteration(1))
-        .subscribe(() => {
-          duration1 = Date.now() - start1;
-
-          of('test')
-            .pipe(moveInEventLoopIteration(2))
-            .subscribe(() => {
-              duration2 = Date.now() - start2;
-
-              expect(duration2).toBeGreaterThan(duration1);
-              done();
-            });
-        });
-    });
-  });
 });
